# swagger-ui/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "the purpose of this application is create metrics system",
    "title": "true-tickets challenge",
    "version": "0.0.1"
  },
  "host": "localhost:9091",
  "basePath": "/",
  "paths": {
    "/metric/{key}": {
      "post": {
        "description": "Endpoint to add a new metric",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["metrics"],
        "operationId": "AddMetricRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Key",
            "description": "The metric key",
            "name": "key",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Value",
            "description": "The metric value",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/MetricInput"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AddMetricResponse"
          }
        }
      }
    },
    "/metric/{key}/sum": {
      "get": {
        "description": "Endpoint to obtain the sum for a given metric",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["metrics"],
        "operationId": "SumMetricRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Key",
            "description": "The metric key",
            "name": "key",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SumMetricResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "MetricInput": {
      "description": "Represents the metric value",
      "type": "object",
      "properties": {
        "value": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "value"
        }
      },
      "x-go-package": "github.com/true-tickets-challenge/model"
    }
  },
  "responses": {
    "AddMetricResponse": {
      "description": "AddMetricResponse just notify if there is any error adding the new metric"
    },
    "SumMetricResponse": {
      "description": "SumMetricResponse retrieves the sum of a given metric in the last 60 minutes",
      "headers": {
        "value": {
          "type": "integer",
          "format": "int64",
          "description": "reprensents the sum of the metric"
        }
      }
    }
  }
}